refactor: Remove landing-page
