+++ apps/explorer/components/SNIHeader.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="navbar bg-base-content px-10 py-5 text-white">
     <div class="navbar-start">
-      <NuxtLink to="/search"
+      <NuxtLink to="/"
         ><img class="w-8 md:w-12" src="~/assets/images/sni-logo.svg"
       /></NuxtLink>
     </div>
