--- apps/explorer/pages/search.vue
@@ -1,95 +1,0 @@
-<template>
-  <div>
-    <div class="lg:px-24">
-      <div v-if="error" class="mx-24 mt-2 h-screen">
-        <SNIAlert
-          class="delay-700 duration-500 ease-in-out"
-          :class="{
-            'opacity-0': !isActive
-          }"
-        />
-      </div>
-      <div v-else>
-        <SNIFilter @search-filter="handleSearchFilter"></SNIFilter>
-        <SNITableMobile
-          :results="result ? result : souls"
-          class="mx-6 py-12 lg:hidden"
-        />
-        <SNITable
-          :results="result ? result : souls"
-          class="hidden lg:block lg:h-screen lg:w-full lg:py-12"
-        />
-      </div>
-    </div>
-  </div>
-</template>
-<script setup lang="ts">
-import { Soul } from '~~/types/soul'
-
-useGqlCors({ credentials: 'same-origin' })
-const { data, error } = await useAsyncGql('sniList', { sniId: '1' })
-let details: Soul[] = []
-const isActive = ref(false)
-let souls = ref([] as Soul[])
-let result = $ref(null as Soul[])
-
-if (error.value) {
-  isActive.value = true
-  displayErrorMessage()
-} else {
-  details = data.value.snis as Soul[]
-  souls = useSouls(details as Soul[])
-}
-
-function displayErrorMessage() {
-  isActive.value = true
-
-  setTimeout(() => {
-    isActive.value = false
-  }, 5000)
-}
-
-function handleSearchFilter(s) {
-  result = s
-}
-</script>
-<style lang="scss">
-@import '@vuepic/vue-datepicker/src/VueDatePicker/style/main.scss';
-// datepicker customization
-.dp__theme_dark {
-  --dp-background-color: #5d5656;
-  --dp-text-color: #ffffff;
-  --dp-hover-color: #484848;
-  --dp-hover-text-color: #ffffff;
-  --dp-hover-icon-color: #959595;
-  --dp-primary-color: #5c7f67;
-  --dp-primary-text-color: #ffffff;
-  --dp-secondary-color: #ecf4e7;
-  --dp-border-color: #2d2d2d;
-  --dp-menu-border-color: #2d2d2d;
-  --dp-border-color-hover: #ffffff;
-  --dp-disabled-color: #737373;
-  --dp-scroll-bar-background: #212121;
-  --dp-scroll-bar-color: #484848;
-  --dp-success-color: #aae6b9;
-  --dp-success-color-disabled: #5c7f67;
-  --dp-icon-color: #ffffff;
-  --dp-danger-color: #e53935;
-  --dp-highlight-color: rgba(0, 92, 178, 0.2);
-
-  .dp__input_wrap {
-    .dp__input {
-      border-radius: 50px;
-      border: none;
-      padding-top: 0.75rem;
-      padding-bottom: 0.75rem;
-      &::placeholder {
-        color: #ffffff;
-        opacity: 1;
-        font-weight: 600;
-        font-size: 0.875rem;
-      }
-    }
-  }
-}
-</style>
